Move MetaPipeEntity registrations to separated class
